+++ C#/Phase08/ElasticFinderConsoleApp/CustomException/RequestNotSentException.cs
@@ -9,7 +9,5 @@ namespace ElasticFinderConsoleApp.CustomException
         public RequestNotSentException(string message) : base(message)
         {
         }
-
-        
     }
 }
