+++ C#/Phase08/ElasticFinderConsoleApp/Model/Document.cs
@@ -1,11 +1,11 @@
 using System;
 
-namespace ElasticFinderConsoleApp
+namespace ElasticFinderConsoleApp.Model
 {
     public class Document
     {
-        public string Name{get; set;}
-        public string Content{get; set;}
+        public string Name { get; set; }
+        public string Content { get; set; }
 
         public override bool Equals(object obj)
         {
