+++ C#/Phase08/ElasticFinderConsoleApp/Program.cs
@@ -1,5 +1,5 @@
 ﻿using System.Threading.Tasks;
-using Nest;
+using ElasticFinderConsoleApp.View;
 
 namespace ElasticFinderConsoleApp
 {
